Presentational and Container Components

+++ src/components/DishdetailComponent.js
@@ -15,7 +15,7 @@ class Dishdetail  extends Component{
                     &nbsp;
                     {new Intl.DateTimeFormat('en-US', {
                             year: 'numeric',
-                            month: 'long',
+                            month: 'short',
                             day: '2-digit'
                         }).format(new Date(comment.date))}
                     </p>
@@ -44,11 +44,11 @@ class Dishdetail  extends Component{
                     </CardBody>
                     </Card>
                 </div>  
-            )
+                 )
         }
         else{
             return (<div></div>)
-        }
+            }
     }
     render() {
         const dish = this.props.dish
@@ -58,10 +58,10 @@ class Dishdetail  extends Component{
         const dishItem = this.renderDish(dish)
         const commentItem = this.renderComments(dish.comments)
         return (
-             <div className="row">
-                 {dishItem}
+            <div className="row">
+                  {dishItem}
                  {commentItem}
-             </div>
+            </div>
         );
     }
 }
